feat: include maintenance days in login streak calculation
- protect streaks during planned, active, and ended maintenances
- ignore canceled maintenances
- replace recursive date expansion with MariaDB sequences
- use long for calculated play time
- invalidate cached stats after maintenance changes
- remove unused user_stats table

+++ src/main/java/de/murmelmeister/murmelapi/user/stats/UserStats.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 public interface UserStats {
     int userId();
 
-    int playTime();
+    long playTime();
 
     int dailyStreak();
 
@@ -18,7 +18,7 @@ public interface UserStats {
 
     long currentPlayTime();
 
-    static @NotNull UserStats of(int userId, int playTime, int dailyStreak, @Nullable LocalDateTime lastSeenAt, boolean isOnline) {
+    static @NotNull UserStats of(int userId, long playTime, int dailyStreak, @Nullable LocalDateTime lastSeenAt, boolean isOnline) {
         return new UserStatsImpl(userId, playTime, dailyStreak, lastSeenAt, isOnline);
     }
 }
